feat: `mutate` and `parse` returns a `Either` instead of `Try` (#172)
Returning an `Either` is considered cleaner than returning a `Try` because it is more explicit. Where a `Try` can mean '_any_ exception' (even `OutOfMemoryException` or unrelated things like `ConnectException`), an `Either` has a more strict domain of failures. Additionally, a `String` failure result makes it easier to handle and return to the user.

Co-authored-by: Nhat <nhaajt@gmail.com>