feat: add replacement field to Mutant (#274)
Adds a new field `replacement` to the Mutant class. This field contains the actually newly replaced part of the pattern, corresponding to the `location`.

Reason for adding this is because there is a mismatch between `pattern` and `location`. So for accurate location reporting, we need the part of the `location` that has been changed.